change type of welcome logo string array data

+++ include/welcome_logo.h
@@ -1,8 +1,6 @@
 #pragma once
 
-#include <stdio.h>
-
 extern const char *welcome_logo[];
 
-extern const size_t WELCOME_LOGO_ROWS;
-extern const size_t WELCOME_LOGO_COLS;
+extern const int WELCOME_LOGO_ROWS;
+extern const int WELCOME_LOGO_COLS;
